refactor: 코드 품질 개선 - 데드코드 제거 및 SSR 안전성 강화
- HotArticle: 모바일에서 쿼리 비활성화 시 스피너 렌더링 방지를 위한 조기 반환 추가
- MobileArticleDetailFooter: JS isMobile 분기 제거, SCSS 반응형 스타일에 위임하여 SSR 일관성 확보
- 공지 목록 Pagination: 데드코드 null 분기 제거, total_count → total_page로 수정
- ProfilePage: 시간표 그리드 컴포넌트를 components/ProfilePage/TimetablePreview로 분리
