+++ src/components/ProfilePage/TimetablePreview/TimetablePreview.module.scss
@@ -1,0 +1,99 @@
+.timetable {
+  &__gridBoard {
+    display: grid;
+    grid-template-columns: 30px repeat(5, minmax(0, 1fr));
+    grid-template-rows: 16px 292px;
+    box-sizing: border-box;
+    width: 100%;
+    height: 316px;
+    padding-top: 8px;
+    overflow: hidden;
+    background: #fff;
+  }
+
+  &__corner {
+    grid-column: 1;
+    grid-row: 1;
+    border-bottom: 1px solid #f5f5f5;
+  }
+
+  &__day {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-bottom: 1px solid #f5f5f5;
+    color: #727272;
+    font-size: 12px;
+    font-weight: 500;
+    line-height: 1.6;
+  }
+
+  &__hours {
+    display: grid;
+    grid-column: 1;
+    grid-row: 2;
+    grid-template-rows: repeat(10, 1fr);
+    border-right: 1px solid #f5f5f5;
+    color: #727272;
+    font-size: 12px;
+    font-weight: 500;
+    line-height: 1.6;
+
+    span {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      border-bottom: 1px solid #f5f5f5;
+    }
+  }
+
+  &__grid {
+    display: grid;
+    grid-column: 2 / 7;
+    grid-row: 2;
+    grid-template-columns: repeat(5, minmax(0, 1fr));
+    grid-template-rows: repeat(20, 1fr);
+    background:
+      repeating-linear-gradient(
+        to bottom,
+        transparent 0,
+        transparent calc(5% - 1px),
+        #f5f5f5 calc(5% - 1px),
+        #f5f5f5 5%
+      ),
+      repeating-linear-gradient(
+        to right,
+        transparent 0,
+        transparent calc(20% - 1px),
+        #f5f5f5 calc(20% - 1px),
+        #f5f5f5 20%
+      );
+  }
+
+  &__block {
+    z-index: 1;
+    display: flex;
+    flex-direction: column;
+    gap: 2px;
+    box-sizing: border-box;
+    min-width: 0;
+    padding: 4px;
+    overflow: hidden;
+    color: #0b0b0d;
+
+    span {
+      overflow: hidden;
+      font-size: 10px;
+      font-weight: 500;
+      line-height: 1.6;
+    }
+
+    small {
+      overflow: hidden;
+      font-size: 8px;
+      line-height: 1.6;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+  }
+}
